refactor: extract native server process shell
