+++ crates/jazz-server/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "jazz-server"
+version = "0.1.0"
+edition = "2024"
+
+[features]
+default = []
+otel = ["dep:jazz-otel", "dep:opentelemetry"]
+
+[dependencies]
+jazz = { workspace = true, default-features = false, features = ["server"] }
+jazz-otel = { workspace = true, optional = true }
+axum = { version = "0.8.9", default-features = false, features = ["http1", "tokio"] }
+opentelemetry = { version = "0.31", optional = true }
+tokio = { version = "1", features = ["net", "rt", "signal", "sync", "time"] }
+tracing = "0.1"
